backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
motor==3.7.1
pymongo==4.12.1
python-dotenv==1.1.0
